fix: build of docs and cadence
