+++ src/cadence/package.json
@@ -9,7 +9,7 @@
     "coverage": "vitest run --coverage",
     "lint": "eslint src",
     "lint:fix": "eslint src --fix",
-    "generate:playground": "tsc --project tsconfig.playground-declaration.json && spack"
+    "generate:playground": "tsc --project tsconfig.json && spack"
   },
   "keywords": [
     "music",
@@ -19,14 +19,14 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@stylistic/eslint-plugin": "^1.5.1",
-    "@swc/cli": "^0.1.62",
-    "@swc/core": "^1.3.95",
-    "@typescript-eslint/eslint-plugin": "^6.11.0",
-    "@typescript-eslint/parser": "^6.11.0",
+    "@stylistic/eslint-plugin": "^1.5.3",
+    "@swc/cli": "^0.1.63",
+    "@swc/core": "^1.3.102",
+    "@typescript-eslint/eslint-plugin": "^6.18.1",
+    "@typescript-eslint/parser": "^6.18.1",
     "@vitest/coverage-v8": "^1.1.3",
-    "eslint": "^8.53.0",
-    "typescript": "^5.2.2",
-    "vitest": "^1.0.0"
+    "eslint": "^8.56.0",
+    "typescript": "^5.3.3",
+    "vitest": "^1.1.3"
   }
 }
