+++ src/cadence/src/AudioLoader.ts
@@ -1,7 +1,6 @@
 import { Loop, Song } from './types'
-import { LoadedLoop, LoadedLoopMock } from './LoadedLoop'
+import { LoadedLoop } from './LoadedLoop'
 import { Timer } from './Timer'
-import { vi } from 'vitest'
 
 export class AudioLoader {
   private loadedLoop: LoadedLoop[] = []
@@ -48,10 +47,3 @@ export async function decodeAudioData(arrayBuffer: ArrayBuffer): Promise<AudioBu
   const audioContext = new AudioContext()
   return await audioContext.decodeAudioData(arrayBuffer)
 }
-
-export const AudioLoaderMock = vi.fn(() => ({
-  loadSong: vi.fn(),
-  getLoops: vi.fn(() => [
-    new LoadedLoopMock(),
-  ]),
-}))
