+++ src/cadence/tsconfig.json
@@ -1,23 +1,18 @@
 {
   "include": [
-    "./src/**/*.ts"
+    "src/**/*.ts",
   ],
   "exclude": [
-    "node_modules"
+    "node_module",
+    "src/**/*.spec.ts",
+    "src/**/*.test.ts"
   ],
   "compilerOptions": {
-    "target": "ES2020"
-    /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019', 'ES2020', or 'ESNEXT'. */,
-    "declaration": true
-    /* Generates corresponding '.d.ts' file. */,
-    "declarationMap": true
-    /* Generates a sourcemap for each corresponding '.d.ts' file. */,
+    "target": "ES2020",
     "preserveSymlinks": true,
-    "sourceMap": true
-    /* Generates corresponding '.map' file. */,
+    "sourceMap": true,
     "downlevelIteration": true
     /* Provide full support for iterables in 'for-of', spread, and destructuring when targeting 'ES5' or 'ES3'. */,
-    "isolatedModules": false,
     /* Transpile each file as a separate module (similar to 'ts.transpileModule'). */
 
     /* Strict Type-Checking Options */
@@ -58,7 +53,13 @@
     "esModuleInterop": true
     /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */,
     /* Advanced Options */
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
     /* Disallow inconsistently-cased references to the same file. */
+    "outFile": "../playground/public/cadence.d.ts",
+    "module": "system",
+    "emitDeclarationOnly": true,
+    "declaration": true,
+    "isolatedModules": false,
+    "declarationMap": false
   }
 }
