--- src/cadence/tsconfig.playground-declaration.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "exclude": [
-    "src/**/*.spec.ts",
-    "src/**/*.test.ts"
-  ],
-  "compilerOptions": {
-    "outFile": "../playground/public/cadence.d.ts",
-    "module": "system",
-    "emitDeclarationOnly": true,
-    "declaration": true,
-    "isolatedModules": false,
-    "declarationMap": false
-  }
-}
